Scrap state-comparing, doesn't work in practice; cleanup
Comparing states was too volatile. It was nice in theory, providing an
interface for "popup" events, as well as making animating new spells/
items easy. In its stead:

For "popup" events, different webpages will be created that render
different things. For example, a dedicated "tier status" webpage will
render the status of the tiers for players based on the state. A
streamer can then switch to that view when that information is
relevant.

For animating spells/items and such, vnode lifecycle events can be
used. This is ex. when a new "hero component" gets created by mithril
(or react or whatever) it can be animated in. Same for spells and
items.