WIP: Better server, initial client structure
This makes the server accept more arguments and gives it
interchangeable arguments between config file and command-line.

This makes the server both send the state and now a config object as
well. This config object can be changed by the client if it sends a
corresponding object, which updates the config for all listening
clients.

This adds a base for the client. It uses a special webpack
configuration that builds all "views" in a views/ directory. These
views can make use of components and utils in a common/ directory.
These views produce local html files that can be used by OBS to render
the in-game data. The UI framework used is mithril.js: React-like, but
much more lightweight.

Eventually this client package will be published on npm, from there
anyone can install it as a dependency and develop their own views by
extending its webpack config, being able to inherit/use the common
components as well as override sass variables with their own sass.

Coming soon:
 - Finishing the basic "camo" client (an unopinionated client that
   blends in with the WC3 UI)
 - Publishing the client package on npm
 - Creating a deploy script that runs pyinstaller, extract_icons.py and
   webpack to produce a distribution package that can be released on
   GitHub