CamoHeroOverviewComponent: Camouflage the level/charges indicators
This may look worse than before depending on stream quality, and they
get really tiny on the items. Can be reverted but is cool in that now
there is no arbitrary choice made whatsoever when it comes to the look.

May also be confusing on the hero icons, since in-game it represents
unused skill points.