server: Stability on new clients
This also bumps the server version to 0.6.0.

I am a websocket and async(io) noob. I think this makes sense. It also
gets rid of that annoying exception message, so, closes #1

I feel like there could be a way to keep only a single instance of the
game but still synced across clients, meaning o(1) memory footprint,
instead of increasing with each new client. But that sounds advanced,
so, TBD.