client: Attempt to solve low contrast player colors
This is a bit of a hack solution but I struggled to find another
better before deciding on this. It takes colors that are below a
certain YIQ threshold and lightens them up. This means the player color
is being lied about, but, unless both players are using colors of
similar hues (e.g. green and emerald, gray and coal), it's easy to
deduce which player is which.