client: Edit less player colors
This decreases the threshold, meaning colors blue, emerald and
dark-green are no longer edited on the resource panel, leaving only
purple, brown, maroon, navy and coal being edited. The intention had
been mainly to address the color purple, the worst offender, and
definitely not to change one of the two basic colors, blue.