server: Use single game instance for all clients
I'm not sure if this is any helpful at all, or if it just meddles with
general websocket behavior. It keeps track of connected clients then
has its own loop to send the game state to those clients. This is
probably bad because iterating and sending the state to all clients is
done synchronously, though that may be the case for websockets already.
Again, I'm a noob at this stuff.

The benefits of this is less memory usage since not every client is
reading the state and sending it individually. Probably negligible
gains, however.

Leaving these changes on a separate branch for now to be put on ice.