server: Be not dumb
Well, this alleviates most of the worries mentioned in the last commit.

Also, while sending the state immediately is cute, it's probably
unnecessary. Just send it when the update hits. It's only a problem if
for example you shutdown the source in OBS and switch to the scene, it
can take up to 2 seconds to "pop in." But it would always "pop in"
regardless, just a consistent couple of milliseconds vs up to 2
seconds. So if you're switching scenes a lot you'd probably not use
that option anyway.