server: Search for players instead of assuming the first N are valid
This is how I originally intended to do it until I noticed the new WC3
client forces players to be in order, i.e. you cannot leave gaps.
However, it turns out Netease does not provide this convenience, so it
does have to search for players after all. It's ugly to hardcode the
player limit, but I don't like the idea of a while loop either. It was
also much prettier to not have to close and re-open the handles every
new game (since almost all games are gonna be 1v1), but what are you
gonna do.

CC #2

This also bumps the server version to 0.8.2.