Add support for Warcraft III: Reforged 1.32
This is a rudimentary implementation of the changes made to the API and
graphics in Reforged while not overhauling any of the current setup.

This was intended to be much better, see #10. While I'm not able to
commit more time to this project, this will have to do.

The server now writes to the shared memory file when the program starts
in order to start the refresh rate.

A new and very simple theme has been added which uses Reforged resource
panels and icons.

Views are now prefixed with hd- or sd- depending on the theme they use.

extract_icons.py now understands the new structure of icons and
extracts the two sets of icons: HD and SD.